chore(deps): update dependencies to latest (#157)
* chore(deps): update dependencies to latest

* chore(deps): tokio back to 0.2

* chore(deps): update tiny-bip39 to 0.8

+++ src/signing/env_mnemonic.rs
@@ -22,7 +22,7 @@ use bee_signing_ext::{
 
 const MNEMONIC_ENV_KEY: &str = "IOTA_WALLET_MNEMONIC";
 const MNEMONIC_PASSWORD_ENV_KEY: &str = "IOTA_WALLET_MNEMONIC_PASSWORD";
-const PBKDF2_ROUNDS: usize = 2048;
+const PBKDF2_ROUNDS: u32 = 2048;
 const PBKDF2_BYTES: usize = 32; // 64 for secp256k1 , 32 for ed25
 
 /// PBKDF2 helper, used to generate [`Seed`][Seed] from [`Mnemonic`][Mnemonic]
